requirements.txt
flask>=3.0.0
reportlab>=4.0.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth>=2.0.0
beautifulsoup4>=4.12.0
openai>=1.0.0
